+++ sdk/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@invariant-labs/a0-sdk",
-  "version": "0.1.20",
+  "version": "0.1.21",
   "collaborators": [
     "Invariant Labs"
   ],
