+++ sdk/src/abis/invariant.ts
@@ -1,7 +1,7 @@
 export const abi = `
 {
   "source": {
-    "hash": "0x964df33de63651de6fba4f877e4cc7ef5d919d7fb2df1b288c123b0d07fefbc0",
+    "hash": "0x43a0f42854fe3a45c884abe00645ce636c2b7d2bfb54049baf085210f6a29019",
     "language": "ink! 4.3.0",
     "compiler": "rustc 1.74.0",
     "build_info": {
