+++ sdk/src/consts.ts
@@ -1,5 +1,4 @@
 import {
-  FeeTier,
   getChunkSize,
   getFeeGrowthDenominator,
   getFeeGrowthScale,
@@ -20,7 +19,6 @@ import {
   getTokenAmountDenominator,
   getTokenAmountScale
 } from '@invariant-labs/a0-sdk-wasm/invariant_a0_wasm.js'
-import { calculateFeeTierWithLinearRatio } from './utils'
 
 export const MAX_REF_TIME = 259058343000
 export const DEFAULT_REF_TIME = 1250000000000
@@ -36,15 +34,6 @@ export const TESTNET_INVARIANT_ADDRESS = '5Fb1ihkMNjPY18qeeKFXUL3GohG1C59JdbfchY
 export const TESTNET_BTC_ADDRESS = '5FEE8ptrT6387MYHqYmyB8ChWfkEsGEDpTMDpwUh4FCYGyCi'
 export const TESTNET_ETH_ADDRESS = '5FmDoQPFS5qPMkSumdvVVekiTpsKVmL9E5DHxHEUXCdHFdYy'
 export const TESTNET_USDC_ADDRESS = '5EjKBBJMLE9R2HsXKJRw2CCMZW2q48Ps5bVAQqzsxyhH9jU5'
-
-export const FEE_TIERS: FeeTier[] = [
-  calculateFeeTierWithLinearRatio(1n),
-  calculateFeeTierWithLinearRatio(2n),
-  calculateFeeTierWithLinearRatio(5n),
-  calculateFeeTierWithLinearRatio(10n),
-  calculateFeeTierWithLinearRatio(30n),
-  calculateFeeTierWithLinearRatio(100n)
-]
 
 export const FEE_GROWTH_DENOMINATOR = getFeeGrowthDenominator()
 export const FIXED_POINT_DENOMINATOR = getFixedPointDenominator()
